refactor: Removing the final weak form references and adding a link to functional from the main doxygen page
