+++ src/serac/physics/utilities/functional/boundary_integral.hpp
@@ -7,17 +7,17 @@
 /**
  * @file integral.hpp
  *
- * @brief This file contains the Integral core of the weak form
+ * @brief This file contains the Integral core of the functional
  */
 #pragma once
 
 #include "mfem.hpp"
 #include "mfem/linalg/dtensor.hpp"
 
-#include "serac/physics/utilities/variational_form/tensor.hpp"
-#include "serac/physics/utilities/variational_form/quadrature.hpp"
-#include "serac/physics/utilities/variational_form/finite_element.hpp"
-#include "serac/physics/utilities/variational_form/tuple_arithmetic.hpp"
+#include "serac/physics/utilities/functional/tensor.hpp"
+#include "serac/physics/utilities/functional/quadrature.hpp"
+#include "serac/physics/utilities/functional/finite_element.hpp"
+#include "serac/physics/utilities/functional/tuple_arithmetic.hpp"
 
 namespace serac {
